Made a base class for opmon publishing
