+++ python/opmonlib/publisher.py
@@ -6,6 +6,7 @@ from google.protobuf.message import Message as Msg
 from google.protobuf.timestamp_pb2 import Timestamp
 
 from opmonlib.conf import OpMonConf
+from opmonlib.publisher_base import opmon_publisher_base
 from opmonlib.utils import (
     LoggingFormatter,
     extract_key,
@@ -19,7 +20,7 @@ from opmonlib.utils import (
 )
 
 
-class OpMonPublisher:
+class OpMonPublisher(opmon_publisher_base):
     """Publish operational monitoring metrics to file or stream."""
 
     def __init__(
