+++ python/opmonlib/publisher_base.py
@@ -1,0 +1,139 @@
+import logging
+import sys
+from abc import ABC, abstractmethod
+
+from google.protobuf.descriptor import FieldDescriptor
+from google.protobuf.message import Message as Msg
+from google.protobuf.timestamp_pb2 import Timestamp
+
+from opmonlib.opmon_entry_pb2 import OpMonEntry, OpMonId, OpMonValue
+
+
+class OpMonPublisherBase(ABC):
+    """Base class for OpMon publishers."""
+
+    @abstractmethod
+    def __init__(self) -> None:
+        """Construct the publisher."""
+        pass
+
+    @abstractmethod
+    def publish(
+        self,
+        session: str,
+        application: str,
+        message: Msg,
+        custom_origin: dict[str, str] | None = None,
+        substructure: list[str] | None = None,
+        level: int | str | None = None,
+    ) -> None:
+        """Publish an OpMonEntry to the relevant location."""
+        pass
+
+    def extract_topic(self) -> str:
+        """Extract the target topic from the message."""
+        self.check_publisher()
+        return self.publisher.default_topic
+
+    def extract_key(self, opmon_entry: OpMonEntry) -> str:
+        """Extract  the key from the OpMonEntry."""
+        self.check_publisher()
+        key = str(opmon_entry.origin.session)
+        if opmon_entry.origin.application != "":
+            key += "." + opmon_entry.origin.application
+        for substructure_id in opmon_entry.origin.substructure:
+            key += "." + substructure_id
+        key += "/" + str(opmon_entry.measurement)
+        return key
+
+    def check_publisher(self) -> None:
+        """Validate that the publisher has a valid method to send messages."""
+        if not self.publisher.opmon_producer:
+            missing_producer_err_str = (
+                "Improperly initialized OpMonProducer used, nothing will be published."
+            )
+            self.log.error(missing_producer_err_str)
+            sys.exit(1)
+        return
+
+    def make_origin(self, session: str, app: str) -> OpMonId:
+        """Construct and return the OpMonId."""
+        return OpMonId(session=session, application=app)
+
+    def make_data(self, message: Msg, top_block: str = "") -> dict:
+        """Map each message entry to the correct data type."""
+        message_dict = {}
+        for name, descriptor in message.DESCRIPTOR.fields_by_name.items():
+            if descriptor.label == FieldDescriptor.LABEL_REPEATED:
+                continue  # Repeated values not supported in influxdb
+            if descriptor.cpp_type == FieldDescriptor.CPPTYPE_MESSAGE:
+                top_block += name + "."
+                message_dict = message_dict | self.make_data(
+                    getattr(message, name), top_block
+                )
+            else:
+                message_dict[top_block + name] = self.to_map(
+                    value=getattr(message, name), field_type=descriptor.cpp_type
+                )
+        return message_dict
+
+    def validate_custom_origin(
+        self, custom_origin: dict[str, str] | None = None
+    ) -> dict[str, str]:
+        """Validate that each custom_origin entry is a str."""
+        if custom_origin is None:
+            return None
+        for key, value in custom_origin.items():
+            if not isinstance(value, str):
+                try:
+                    custom_origin[key] = str(value)
+                except TypeError:
+                    msg = "%s is not a string and cannot be converted to one.", key
+                    raise TypeError(msg) from None
+        return custom_origin
+
+    def to_map(self, value: int | float | bool | str, field_type: int) -> OpMonValue:
+        """Map the data entry to the correct protobuf format."""
+        formatted_opmonvalue = OpMonValue()
+        match field_type:
+            case FieldDescriptor.CPPTYPE_INT32:
+                formatted_opmonvalue.int4_value = value
+            case FieldDescriptor.CPPTYPE_INT64:
+                formatted_opmonvalue.int8_value = value
+            case FieldDescriptor.CPPTYPE_UINT32:
+                formatted_opmonvalue.uint4_value = value
+            case FieldDescriptor.CPPTYPE_UINT64:
+                formatted_opmonvalue.uint8_value = value
+            case FieldDescriptor.CPPTYPE_DOUBLE:
+                formatted_opmonvalue.double_value = value
+            case FieldDescriptor.CPPTYPE_FLOAT:
+                formatted_opmonvalue.float_value = value
+            case FieldDescriptor.CPPTYPE_BOOL:
+                formatted_opmonvalue.boolean_value = value
+            case FieldDescriptor.CPPTYPE_STRING:
+                formatted_opmonvalue.string_value = value
+            # Ignore unknown types.
+        return formatted_opmonvalue
+
+    def to_entry(
+        self,
+        session: str,
+        application: str,
+        message: Msg,
+        custom_origin: dict[str, str] | None,
+        substructure: list[str] | None,
+        t: Timestamp,
+        log: logging.Logger | None = None,
+    ) -> OpMonEntry:
+        """Pack all the data that needs to be published to an OpMonEntry."""
+        o = OpMonEntry(
+            time=t,
+            origin=self.make_origin(session, application),
+            custom_origin=self.validate_custom_origin(custom_origin),
+            measurement=message.DESCRIPTOR.full_name,
+            data=self.make_data(message),
+        )
+        if log:
+            log.error(t)
+            log.error(o.time)
+        return o
